+++ test/edu/umass/cs/gnsclient/client/singletests/CreateGuidBatchTestWithPublicKeys.java
@@ -91,7 +91,7 @@ public class CreateGuidBatchTestWithPublicKeys {
       aliases.add("testGUID" + RandomString.randomString(6));
     }
     String result = null;
-    int oldTimeout = client.getReadTimeout();
+    long oldTimeout = client.getReadTimeout();
     try {
       client.setReadTimeout(15 * 1000); // 30 seconds
       result = client.guidBatchCreate(masterGuid, aliases);
